// src/user/dto/enum/enum-user.dto.ts
export enum DefaultOrderUserDto {
    user_id = 1,
    firstName = 'MEC_Name',
    lastName = 'MEC_Surname',
    email = 'test@mec.com',
    country = 'Netherlands',
    phoneNumber = '+000 1234567',
    city = 'Amsterdam',
    zipCode = 'AA-1234',
    streetName = 'some street',
    houseNumber = 102,
    company = 'SIA SomeCompany',
    tax_id = '12110098765'
}